flask_app/requirements.txt
Flask==1.1.4
Flask-SQLAlchemy==2.4.4
pymysql==1.0.2
sqlalchemy==1.3.23
Werkzeug==1.0.1
MarkupSafe==1.1.1
Jinja2==2.11.3
DBUtils==3.1.0
python-dotenv==1.0.1
